moved schema endpoint functions and tests into their own files
