+++ apps/schematic/api/.openapi-generator-ignore
@@ -33,4 +33,5 @@ schematic_api/__main__.py
 setup.py
 test-requirements.txt
 schematic_api/test/test_storage_controller.py
+schematic_api/test/test_schema_controller.py
 tox.ini
