+++ apps/schematic/api/.openapi-generator/FILES
@@ -1,6 +1,7 @@
 AUTHORS.md
 schematic_api/__init__.py
 schematic_api/controllers/__init__.py
+schematic_api/controllers/schema_controller.py
 schematic_api/controllers/security_controller_.py
 schematic_api/controllers/storage_controller.py
 schematic_api/encoder.py
@@ -20,5 +21,6 @@ schematic_api/models/manifests_page_all_of.py
 schematic_api/models/page_metadata.py
 schematic_api/openapi/openapi.yaml
 schematic_api/test/__init__.py
+schematic_api/test/test_schema_controller.py
 schematic_api/typing_utils.py
 schematic_api/util.py
