+++ apps/schematic/api/schematic_api/controllers/utils.py
@@ -1,0 +1,49 @@
+from typing import Callable, Union, Any
+
+from synapseclient.core.exceptions import (  # type: ignore
+    SynapseNoCredentialsError,
+    SynapseAuthenticationError,
+)
+from schematic.exceptions import AccessCredentialsError  # type: ignore
+
+from schematic_api.models.basic_error import BasicError
+
+
+def handle_exceptions(endpoint_function: Callable) -> Callable:
+    """
+    This is designed to be used as a decorator for endpoint functions.
+    The endpoint function is called in a try block, and then various
+      Synapse and Schematic exceptions are handled and returned as the
+      BasicError object.
+
+    Args:
+        f (Callable): A function that calls the input function
+    """
+
+    def func(*args: Any, **kwargs: Any) -> tuple[Union[Any, BasicError], int]:
+        try:
+            return endpoint_function(*args, **kwargs)
+
+        except SynapseNoCredentialsError as error:
+            status = 401
+            res = BasicError(
+                "Missing or invalid Synapse credentials error", status, str(error)
+            )
+            return res, status
+
+        except SynapseAuthenticationError as error:
+            status = 401
+            res = BasicError("Forbidden Synapse access error", status, str(error))
+            return res, status
+
+        except AccessCredentialsError as error:
+            status = 403
+            res = BasicError("Synapse entity access error", status, str(error))
+            return res, status
+
+        except Exception as error:  # pylint: disable=broad-exception-caught
+            status = 500
+            res = BasicError("Internal error", status, str(error))
+            return res, status
+
+    return func
