+++ apps/schematic/api/schematic_api/test/test_storage_controller_impl.py
@@ -1,4 +1,4 @@
-"""Tests for endpoint functions"""
+"""Tests for storage endpoint functions"""
 
 from unittest.mock import patch
 
@@ -11,12 +11,10 @@ from schematic.exceptions import AccessCredentialsError  # type: ignore
 from schematic_api.models.basic_error import BasicError
 from schematic_api.models.manifests_page import ManifestsPage
 from schematic_api.models.datasets_page import DatasetsPage
-from schematic_api.models.attributes_page import AttributesPage
 import schematic_api.controllers.storage_controller_impl
 from schematic_api.controllers.storage_controller_impl import (
     list_storage_project_datasets,
     list_storage_project_manifests,
-    list_component_attributes,
 )
 
 
@@ -153,36 +151,6 @@ class TestListStorageProjectManifests:
         ):
             result, status = list_storage_project_manifests(
                 project_id="syn1", asset_view_id="syn2", asset_type="synapse"
-            )
-            assert status == 500
-            assert isinstance(result, BasicError)
-
-
-class TestComponentAttributes:
-    """Test case for list_component_attributes"""
-
-    def test_success(self) -> None:
-        """Test for successful result"""
-        with patch.object(
-            schematic_api.controllers.storage_controller_impl,
-            "get_component_attributes",
-            return_value=["attribute1", "attribute2"],
-        ):
-            result, status = list_component_attributes(
-                schema_url="xxx", component_label="label"
-            )
-            assert status == 200
-            assert isinstance(result, AttributesPage)
-
-    def test_internal_error(self) -> None:
-        """Test for 500 result"""
-        with patch.object(
-            schematic_api.controllers.storage_controller_impl,
-            "get_component_attributes",
-            side_effect=TypeError,
-        ):
-            result, status = list_component_attributes(
-                schema_url="xxx", component_label="label"
             )
             assert status == 500
             assert isinstance(result, BasicError)
